fix: a folder had accidentily been made as a file
